+++ Tychaia.ProceduralGeneration/LayerEraseTowns.cs
@@ -18,7 +18,7 @@ namespace Tychaia.ProceduralGeneration
         {
         }
 
-        protected override int[] GenerateDataImpl(int x, int y, int width, int height)
+        protected override int[] GenerateDataImpl(long x, long y, long width, long height)
         {
             if (this.Parents.Length < 2 || this.Parents[0] == null || this.Parents[1] == null)
                 return new int[width * height];
