+++ Tychaia.ProceduralGeneration/LayerInitialVoronoi.cs
@@ -80,7 +80,7 @@ namespace Tychaia.ProceduralGeneration
             this.Result = VoronoiResult.EdgesAndOriginals;
         }
 
-        protected override int[] GenerateDataImpl(int x, int y, int width, int height)
+        protected override int[] GenerateDataImpl(long x, long y, long width, long height)
         {
             int[] data = new int[width * height];
