+++ Tychaia.ProceduralGeneration/LayerTerrainMixdown.cs
@@ -39,7 +39,7 @@ namespace Tychaia.ProceduralGeneration
             this.MaxPerlin = 100;
         }
 
-        protected override int[] GenerateDataImpl(int x, int y, int width, int height)
+        protected override int[] GenerateDataImpl(long x, long y, long width, long height)
         {
             if (this.Parents.Length < 2 || this.Parents[0] == null || this.Parents[1] == null)
                 return new int[width * height];
