+++ Tychaia.ProceduralGeneration/LayerZoomWaterDistance.cs
@@ -58,14 +58,14 @@ namespace Tychaia.ProceduralGeneration
             this.MaxTerrainBinary = 1;
         }
 
-        private int[] GenerateDataIterate(int iter, int x, int y, int width, int height)
+        private int[] GenerateDataIterate(int iter, long x, long y, long width, long height)
         {
             int ox = 4; // Offsets
             int oy = 4;
-            int rw = width / 2 + ox * 2;
-            int rh = height / 2 + oy * 2;
-            int rx = (x < 0 ? (x - 1) / 2 : x / 2) - ox; // Location in the parent
-            int ry = (y < 0 ? (y - 1) / 2 : y / 2) - oy;
+            long rw = width / 2 + ox * 2;
+            long rh = height / 2 + oy * 2;
+            long rx = (x < 0 ? (x - 1) / 2 : x / 2) - ox; // Location in the parent
+            long ry = (y < 0 ? (y - 1) / 2 : y / 2) - oy;
             // For smoothing to work, we need to know the cells that are actually
             // beyond the edge of the center.
             int[] parent = null;
@@ -111,10 +111,10 @@ namespace Tychaia.ProceduralGeneration
             return data;
         }
 
-        private int FindZoomedPoint(int[] parent, int i, int j, int ox, int oy, int x, int y, int rw)
+        private int FindZoomedPoint(int[] parent, long i, long j, long ox, long oy, long x, long y, long rw)
         {
-            int ocx = (x % 2 != 0 && i % 2 != 0 ? (i < 0 ? -1 : 1) : 0);
-            int ocy = (y % 2 != 0 && j % 2 != 0 ? (j < 0 ? -1 : 1) : 0);
+            long ocx = (x % 2 != 0 && i % 2 != 0 ? (i < 0 ? -1 : 1) : 0);
+            long ocy = (y % 2 != 0 && j % 2 != 0 ? (j < 0 ? -1 : 1) : 0);
 
             // Create localized values for immediate surrounding cells.
             int v00 = parent[(i / 2 + ox + ocx - 1) + (j / 2 + oy + ocy - 1) * rw];
@@ -139,7 +139,7 @@ namespace Tychaia.ProceduralGeneration
             return LayerColors.GetTerrainBrushes(1 << this.MaxTerrainBinary);
         }
 
-        protected override int[] GenerateDataImpl(int x, int y, int width, int height)
+        protected override int[] GenerateDataImpl(long x, long y, long width, long height)
         {
             if (this.Iterations > 0)
                 return this.GenerateDataIterate(1, x, y, width, height);
@@ -149,7 +149,7 @@ namespace Tychaia.ProceduralGeneration
                 return this.Parents[0].GenerateData(x, y, width, height);
         }
 
-        private int Smooth(int x, int y, int northValue, int southValue, int westValue, int eastValue, int currentValue, int i, int j, int ox, int oy, int rw, int[] parent)
+        private int Smooth(long x, long y, int northValue, int southValue, int westValue, int eastValue, int currentValue, long i, long j, long ox, long oy, long rw, int[] parent)
         {
             // Parent-based Smoothing
             int selected = 0;
